% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartmusicdata-package.R
\docType{package}
\name{chartmusicdata-package}
\alias{chartmusicdata}
\alias{chartmusicdata-package}
\title{chartmusicdata: Contains Data from chart2000.com and tsort.info}
\description{
Holds data from chart2000.com and tsort.info about chart success of songs and albums.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fjodor/chartmusicdata}
  \item Report bugs at \url{https://github.com/fjodor/chartmusicdata/issues}
}

}
\author{
\strong{Maintainer}: Wolf Riepl \email{wolf.riepl@statistik-dresden.de}

}
\keyword{internal}
