% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartmusicdata.R
\docType{_PACKAGE}
\name{chartmusicdata}
\alias{chartmusicdata}
\title{chartmusicdata: A data package that contains data from https://chart2000.com and https://tsort.info.}
\description{
Datsets in this package: charts, topalbums, topsongs, albums2000, songs2000
}
