% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\docType{data}
\name{songs2000}
\alias{songs2000}
\title{Top 50 songs for every month from Jan 2000 to Jun 2021.}
\format{
A data frame with 12900 rows and 13 variables:
\describe{
\item{year_month}{year and month, character, format "2000-01"}
\item{month}{month allocated to entry, ordered factor, levels from 1 = "Jan" to 12 = "Dec"}
\item{year}{year allocated to entry, numeric, 4 digits}
\item{position}{total rank, numeric from 1 to 50}
\item{artist}{name of artist / band, character}
\item{song}{song title, character}
\item{indicativerevenue}{estimated revenue across whole music chain in thousands of dollars; numeric}
\item{us}{peak position in US in that month, numeric, contains NAs}
\item{uk}{peak position in UK in that month, numeric, contains NAs}
\item{de}{peak position in Germany in that month, numeric, contains NAs}
\item{fr}{peak position in France in that month, numeric, contains NAs}
\item{ca}{peak position in Canada in that month, numeric, contains NAs}
\item{au}{peak position in Australia in that month, numeric, contains NAs}
}
}
\source{
\url{https://chart2000.com/about}
}
\usage{
songs2000
}
\description{
A dataset containing the top 50 songs per month from January 2000 to June 2021
according to charts from the USA, UK, Germany, France, Canada, Australia, Italy and Spain.
}
\examples{
data(songs2000)        # Lazy loading. Data becomes visible as soon as used, e. g. via str(songs2000).

## Access file version number:
attr(songs2000, "version")
str(songs2000)
}
\keyword{datasets}
